Relocating one file to remove similar package names.
I made a util and utils package, dumb oversight by me.  Consolidated into one.
